# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Europe TikTok B2B Video Learning & Packaging Adaptation Assistant",
    "description": "API for managing TikTok B2B video research, inspiration library, daily hot video collection, and content recommendations for flexible packaging factories targeting European markets.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:3002", "description": "Local development" }
  ],
  "paths": {
...
    "/api/inspiration-videos": {
      "get": {
        "summary": "List inspiration videos",
        "tags": ["Inspiration"],
        "parameters": [
          { "name": "platform", "in": "query", "schema": { "type": "string" } },
          { "name": "video_type", "in": "query", "schema": { "type": "string" } },
          { "name": "reference_industry", "in": "query", "schema": { "type": "string" } },
          { "name": "value_level", "in": "query", "schema": { "type": "string" } },
          { "name": "status", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "List of inspiration videos" } }
      },
      "post": {
        "summary": "Add inspiration video",
        "tags": ["Inspiration"],
        "responses": { "200": { "description": "Created" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Set API_AUTH_TOKEN in .env, then pass as Bearer token"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Set API_AUTH_TOKEN in .env, then pass as x-api-key header"
      }
    }
  }
}